Stop using 2-pass build on OS X
This is a follow up CL for r211. No behavioral change expected with this CL.

Seems that we need to teach xcodebuld about the dependency explicitly when the code generators are used in actions. With this CL, build_mozc.py build_tools is no longer required on OS X as well as other supported platforms.

This CL closes Issue 223.

BUG=mozc:223
TEST=no behavioral change expected


git-svn-id: https://mozc.googlecode.com/svn/trunk@214 a6090854-d499-a067-5803-1114d4e51264