Use absolute path to teach the linker about a file location
This is a groundwork for Ninja migration on Linux. No user-visible behavioral changes are expected with this CL.

We have to use the absolute path for a file passed as a linker option script because we cannot assume that the current directory is the same to the source root directory with Ninja.

BUG=mozc:222
TEST=compile and unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@219 a6090854-d499-a067-5803-1114d4e51264