Do not use shell syntax in gyp (part 1)
This is a groundwork for Ninja migration on Linux. No user-visible behavioral changes are expected with this CL.

'$' is reserved for Ninja internal use. We shouldn't rely on it in gyp layer.

With this CL, the environment vairable ANDROID_DEVICES is accessed in run_android_test.py rather than the gyp (and therefore Ninja) layer.

BUG=mozc:222
TEST=compile and unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@220 a6090854-d499-a067-5803-1114d4e51264