Do not use shell syntax in gyp (part 2)
This is a groundwork for Ninja migration on Linux. No user-visible behavioral changes are expected with this CL.

'$' is reserved for Ninja internal use. We shouldn't rely on it in gyp layer.

With this CL, a conditional action regarding symbol removal is handled in a python script rather than the gyp (and therefore Ninja) layer.

BUG=mozc:222
TEST=compile and unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@221 a6090854-d499-a067-5803-1114d4e51264