Use different internal namespace for gflags
Mozc's flags system, which is implemented in base/flags.h, looks quite similar to OSS gflags library.  This is not surprising, because they originate from the same Google internal library actually.  As a result, however, currently OSS Mozc and OSS gflags are using the same name for internal variables, which results in a potential risk of conflicting symbol name during link time when you mix Mozc's code with other projects that also rely on gflags.

To eliminate this potential risk, this CL renames the internal namespace from fL##shorttype to mozc_flags_fL##shorttype.  No user-visible behavioral changes are expected with this CL.

BUG=none
TEST=unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@223 a6090854-d499-a067-5803-1114d4e51264