Simplify test cases for SafeStrToDouble/SafeStrToFloat
This CL simplifies test cases for SafeStrToDouble/SafeStrToFloat so as not to assert that SafeStrToFloat/SafeStrToDouble always fail to parse a hexadecimal number with Visual C++ and NDK. Such assertion is too fragile for future update of toolchains and/or platform libraries.

This CL never changes production code.  No functional changes are expected.

BUG=none
TEST=unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@226 a6090854-d499-a067-5803-1114d4e51264