Strictly check the range of the operands for integer operations performed in NumberUtil
Previously, NumberUtil have checked the result of some integer operations instead of checking the range of operands. This isn't good idea, especially for the really trick cases such as |-INT_MIN|.

This CL enforces strict range check of the operands rather than the result for integer operations performed in NumberUtil.

BUG=none
TEST=unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@227 a6090854-d499-a067-5803-1114d4e51264