Ensure that the IPC client is configured before it is used in ibus-mozc
This CL ensures that ClientInterface::set_client_capability is called before the the IPC client is actually used in ibus-mozc.

This is important because Mozc's IPC client library relies on so-called 2-phase construction. The IPC client must be instantiated by ClientFactory::NewClient(), then should be configured appropriately before it is actually used.

Closes Issue 227.

BUG=mozc:227
TEST=manually done on Linux


git-svn-id: https://mozc.googlecode.com/svn/trunk@234 a6090854-d499-a067-5803-1114d4e51264