Do not pass incomplete types to STL containers
Using STL containers with incomplete types are fragile in practice, and easily results in an undefined behavior.  On the other hand, including generated header files in a public header file is also considered harmful, because it easily results in non-deterministic compile errors unless hidden build dependencies are carefully carefully annotated in GYP file.  Thus this CL isn't great as it should be because it deliberately does the later to resolve the former.  Hopefully in future we will revisit here to stop including generated header files in a public header file.

This is just a build-specific issue thus no user-visible changes are expected.

BUG=none
TEST=unittest


git-svn-id: https://mozc.googlecode.com/svn/trunk@250 a6090854-d499-a067-5803-1114d4e51264