Use Ninja to build Android, NaCl, and Linux Desktop binaries part 3
This CL fixes runtime crash on Android.

'product_dir' feature in GYP seems to be interpreted differently between Make generator and Ninja generator especially regarding relative path resolution.  This difference should have been taken care of in r266.  Otherwise, the APK lacks libmozc.so and Mozc for Android crashes whenever it is being launched.

With this CL, 'product_dir' used to build libmozc.so is specified with an absolute path rather than relative path so that Mozc for Android can be built correctly.

BUG= Issue mozc:222 
TEST=manually tested on Android


git-svn-id: https://mozc.googlecode.com/svn/trunk@268 a6090854-d499-a067-5803-1114d4e51264