Add an all-in-one Dockerfile to build Mozc for Android, NaCl, and Linux desktop (Ubuntu 12.04 edition)
This CL adds an all-in-one Dockerfile where everything is ready to build Mozc for Android, NaCl, and Linux desktop.

To set up or update the Docker container:
  . make-buildenv.sh
To enter into the Docker container:
  . enter-buildenv.sh
To set up the Docker container again from scratch:
  . clobber-make-buildenv.sh

This CL is not supposed to introduce any user-visible change.

Closes Issue 240

BUG= Issue mozc:240 
TEST=compile and unittest in Docker 1.0 running on Ubuntu 14.04


git-svn-id: https://mozc.googlecode.com/svn/trunk@271 a6090854-d499-a067-5803-1114d4e51264