Put mozc_build_tools under output directory
Now we have two different intermediate directories for each target platform.  One is (mozc_src_root)/out_{platform_name} and the other is (mozc_src_root)/mozc_build_tools/{os_name}.  This is not only redundant but also confusing.  Actually this is a long-standing TODO as commented in gyp/common.gyp.

With this CL, mozc_build_tools is created as (mozc_src_root)/out_{platform_name}/mozc_build_tools as it is originally planned.

Note that this is just a clean-up of build process.  Hence no behavior change is intended.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@360 a6090854-d499-a067-5803-1114d4e51264