Stop generating xcode rule to run unittest from IDE
We are taking away from Xcode and relying more on Ninja to deal with the complexity of cross platform development.  With this CL, GYP's 'run_as' rule, which allows developers to run unittest from Xcode, is dropped because no one at Mozc team is using that feature.

No user visible change is intended.

BUG=Issue mozc:247
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@363 a6090854-d499-a067-5803-1114d4e51264