Abstract the directory for intermediate files in GYP
Android build has lots of build configurations.  It would be useful if each configuration can have a different directory for intermediate files.  As a groundwork to do so, this CL introduces a new GYP variable that abstracts the intermediate directory.  Currently this CL 'shared_intermediate_mozc_dir' is defined as a simple alias of 'SHARED_INTERMEDIATE_DIR' though.

This change has nothing to do with the production binaries.  Hence no behavior change is intended.

BUG=none
TEST=compile

git-svn-id: https://mozc.googlecode.com/svn/trunk@409 a6090854-d499-a067-5803-1114d4e51264