Shorten GYP target names
On Windows, some build toolchains such as Ninja or GYP may not be able to deal with a long path name that exceeds 256 chars or so.  To work around this limitation, this CL shortens several GYP target names that tend to be long.

This is completely a build-time change.  Hence no behavior change is intended.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@425 a6090854-d499-a067-5803-1114d4e51264