Support arm64/mips64/x86-64 in Android build
With this CL, arm64/mips64/x86-64 native builds are supported in Android.

To build arm64, run build_mozc.py as follows.

  python build_mozc.py gyp     \
    --target_platform=Android  \
    --android_arch=arm64

Note that this CL also updates the versioning scheme in Android Mozc as follows.

  Previous Version code format:
    0005BBBBBA
    A: ABI (0: Fat, 5: x86, 2: armeabi-v7a, 1:mips)
    B: ANDROID_VERSION_CODE

  New Version code format:
    0005BBBBBA
    A: ABI (0: Fat, 6: x86_64, 5:arm64, 4:mips64, 3: x86, 2: armeabi-v7a, 1:mips)
    B: ANDROID_VERSION_CODE

where ANDROID_VERSION_CODE is equal to the build number defined in mozc_version_template.

Closes Issue 259.

BUG=Issue mozc:259
TEST=compile with Android NDK r10d

git-svn-id: https://mozc.googlecode.com/svn/trunk@439 a6090854-d499-a067-5803-1114d4e51264