For Android and Chrome OS, we need to change the page size of the candidate list from 9 as the default value.
Introduce per-session config mechanism to change the page size of the candidate list

For the traditional desktop-style floating candidate list, in some situations, we may want to reduce its maximum height by changing the page size.  For instance, if we show up-to 5 candidates per page, the candidate list is more likely to be rendered without being cliped by the edge of the screen even on a relatively smaller screen.

With this CL, each mozc client can tell the mozc server to change the page size in a per-session basis.  The default size is kept to be 9 hence no visible change is intended with this CL though.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@447 a6090854-d499-a067-5803-1114d4e51264