Learn content word for real time conversion in the mobile mode
To propagate the segmentation information in mobile use, this CL extends Segment::Candidate::inner_segment_boundary to include content key and value lengths.  This extra information is utilized in UserHistoryPredictor to reconstruct content word.

Motivation:
In mobile mode, it would be useful if content words can be learned.  Here are some examples.

Case 1:
1. Type がっこうに
2. Select 学校に
学校に (which is one segment) is learned but 学校 (content word) is not learned.  Here learning 学校 makes sense for mobile since particles are suggested by zero query suggestion in the mobile mode.

Case 2:
1. Type とうきょうかなごやにいきたい
2. Select 東京か名古屋に行きたい
This is similar to the case 1 but we can learn both 東京 and 名古屋.

Note that this CL should not change anything in the desktop mode.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@449 a6090854-d499-a067-5803-1114d4e51264