KeyboardView to use BufferedDrawable.
This is a series of CLs for full migration from s/w canvas to h/w
canvas.

KeyboardViewBackgroundSurface gets rid of its internal bitmap buffer and
starts to use BufferedDrawable.  By this change the memory footprint
gets improved.

Previously KeyboardViewBackgroundSurface had internal bitmap buffer. On
onDraw() it only copied the buffer onto the canvas.

The buffer was updated in following way:
 1. KeyboardView registered (potentially) updated keys, keyboard and
    metastate to KeyboardViewBackgroundSurface.
 2. KeyboardView aggressive called invalidate() method even if
    invalidation was not really required.
 3. KeyboardView#onDraw() called KeyboardViewBackgroundSurface#update(),
    in which the buffer was updated based on updated keys. If updated
    keys and so on are empty, the buffere was kept as is (so aggressive
    invalidate() would affect nothing).
 4. Registered updated keys, keyboard and metastate were cleared.

By this change, invalidate() is called as less as possible. To do this
KeyboardViewBackgroundSurface now requires precise information about
pressed keys and so on.

BUG=
TEST=compile
REF_BUG=18564674,19326901
REF_CL=83896650,84257769,85964469