Improve Mushroom plug-in compatibility
This improves compatibility for Mushroop plug-ins that use an
intermediate activity.

Many users have reported that Mushroom doesn't work sometimes (or
always).  The cause is that when activity transition gets like
following:
  Mushroom activity
  -> Intermediate activity
  -> Original application activity,
the result is cleared by intermediate activity.

This CL makes the IME keep the result until it is explicitly consumed or
new Mushroom is launched.

BUG=
TEST=
REF_BUG=18938129
REF_CL=84304463