Fix invalid popup icon size
Previously a Drawable for popup/keyicon instance is shared among all the
classes.  Therefore if an instance's bounds is changed for keyicon, the
popup visual is also changed unexpectedly.  For instance you can type
'A' key on QWERTY-KANA keyboard to reproduce this issue.

Now DrawableCache holds separate Drawable in order to avoid this
problem.

BUG=
TEST=compile
REF_BUG=19117626
REF_CL=84601076,85499615