Make the conditions to learn candidates with a punctuation mark more strict
Learning a candidate with a punctuation was originally introduced for the desktop users so that sentence like suggestions such as "いつもお世話になっております" can be learned with punctuation marks like "。".  However, especially in mobile where zero query suggestions are suggested aggressively, the current algorithm is sometimes too aggressive in practice.

To reduce the risk of polluting history entries, this CL imposes some condition on the history candidate to check if it is a sentence or not.  Currently the candidate is handles as "sentence-like" when all the following conditions are met.
- the reading character length is 8 or greater.
- the word ends with Hiragana character.

Closes Issue 275.

BUG=Issue mozc:275
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@501 a6090854-d499-a067-5803-1114d4e51264