Require Visual C++ 2013 to build Mozc for Windows
To reduce the number of actively maintained build configurations, we will discontinue the support of Visual C++ 2010 and 2012.

With this CL, the required Windows Installer version is also bumped from 3.0 to 4.5 because merge modules of Visual C++ 2013 requires Windows Installer 4.5 or later.  This should be OK because Mozc supports Windows Vista SP2 and later, and Windows Vista SP2 was shipped with Windows Installer 4.5.

Basically no user visible change is intended.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@509 a6090854-d499-a067-5803-1114d4e51264