Make error message from DISALLOW_COPY_AND_ASSIGN and DISALLOW_IMPLICIT_CONSTRUCTORS easier to read
Now that lots of C++11 features are available in Mozc build, we can use 'Deleted Functions' to implement DISALLOW_COPY_AND_ASSIGN and DISALLOW_IMPLICIT_CONSTRUCTORS so that error messages can be easier to read.

Note that this is just a code cleanup thanks to C++11.  No behavior change is intended.

BUG=none
TEST=none

git-svn-id: https://mozc.googlecode.com/svn/trunk@542 a6090854-d499-a067-5803-1114d4e51264