Remove the dependency on NodeAllocatorInterface from dictionary module
Node structure is only relevant to lattice structure in converter module, so its dependency should be removed from dictionary module.  To this end, NodeAllocatorData is removed as it is used only in SystemDictionary for reverse lookup.  Having a general storage for it in NodeAllocatorInterface is overkill.  Also, method signatures of DictionaryInterface are changed accordingly.

node_test.cc is removed as it no longer contains useful tests due to the removal of NodeAllocatorData.

This is just an internal clean-up.  No user-visible behavior change is intended.

BUG=none
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@547 a6090854-d499-a067-5803-1114d4e51264