Keep the content of drawableCache when necessary
Previously drawbleCache had been cleared even if given keyboard is the
same as current keyboard. In this case the content is still valid so
drawableCache can be kept as is.

This aggressive cache-clear logic was not so harmful on old Android
OS.  Even if drawableCache is cleared, a Skin, which backs
drawableCache, also has internal cache so drawable resources was not
created again.  However the internal cache in a Skin uses WeakReference,
which new Android OS doesn't protect from GC anymore.  Therefore when a
Keyboard was newly shown on new Android OS, the task creating drawable
resources had always been executed, which takes much time.

This CL eliminates unneeded clearing the content.  By this CL, the
processing time of first onDraw becomes from 120ms to 5ms.

BUG=none
TEST=manually done
REF_BUG=
REF_CL=86665491