Use Clang 3.4 to build host binaries
In a recent update of clang-3.5 package in Ubuntu 14.04, /usr/bin/clang was renamed to /usr/bin/clang-3.5 for some reasons.
https://bugs.launchpad.net/ubuntu/+source/llvm-toolchain-snapshot/+bug/1420570

As a tentative solution, we will use Clang 3.4 to build host binaries for now.

Note that this CL does not change the toolchanins for the target binaries in Android and NaCl builds.  No user visible change is intended anyway.

Closes 286.

BUG=Issue mozc:286
TEST=unittest

git-svn-id: https://mozc.googlecode.com/svn/trunk@553 a6090854-d499-a067-5803-1114d4e51264