Build Mozc on Fedora 21 in Docker.
This is a proof-of-concept CL that demonstrates how to build Mozc on
Fedora 21.  Note that this CL does not intend to officially support
building and running Mozc on Fedora 21.

To set up or update the Docker container:
  . make-buildenv.sh
To enter into the Docker container:
  . enter-buildenv.sh
To set up the Docker container again from scratch:
  . clobber-make-buildenv.sh

BUG=none
TEST=compile and unittest in Docker 1.6.1 running on Ubuntu 14.04.2