Fix UI freeze by ImeUtil::ActivateForCurrentSessio
Looks like the TSF needs to rebuild a certain cache after changing the
input details. Perhaps this rebuiding process is not resilient enough
for a relatively lot of threads.  The freezing window threads are
waiting for MSCTF.AsmCacheReady.<desktop name><session #> event and
later waiting for an IMM critical section that the former threads own.
This continues for about a minute.

If we wait for the cache ready
(MSCTF.AsmCacheReady.<desktopname><session #> event signal which usually
takes a few hundreds msecs) before broadcasting the input langauge
change, the change seems to be applied flawlessly.

BUG=none
TEST=maunally done
REF_BUG=5765783
REF_CL=84600315