Import a subset of WTL 9.0 into third_party/wtl/.
In order to remove dependency on subversion in DEPS file, a snapshot
copy of WTL 9.0 is directly bundled in OSS Mozc repository at the
following location.
  opensource/mozc/third_party/wtl

This CL also corrects the license notice of WTL, which should be
Common Public License for WTL 9.0 that we are currently using.

As a side note, recently WTL upstream switched back to MS-PL again as
http://sourceforge.net/p/wtl/code/617/ but they have not yet made new
stable release.

No behavior change is intended anyway.

BUG=#299
TEST=compile
REF_BUG=23909405,6443612
REF_CL=102720447