Clarify OSX build does not require Ninja yet.
This is a follow up CL for 276fa5c362e68771b840dd5925317e310d6d75ff,
which mistakenly stated that OSX build requires Ninja.

Although we still plan to use Ninja for OSX build as mentioned
in #247, it is not yet done.  Until then, we don't have to ask people
to install Ninja when building Mozc for OSX.

BUG=#247
TEST=unittest
REF_BUG=
REF_CL=