Remove unnecessary '-r' option from 'ln' command.
For this particular case, we do not have to specify '-r' option.
Even if we remove '-r' option here, exactly the same symlink will be
created in practice.

Not relying on '-r' option is convenient since '-r' option does not work
with older versions of 'ln' command.  In particular, that option does
not work on Ubuntu 12.04, which prevents us from enabling Android build
on Travis-CI.

This is just a code clean up in build rules.  No behavior change in
artifacts should occur.

BUG=none
TEST=compile
REF_BUG=none
REF_CL=105034319