Do not convert NSNotFound to an int variable.
As explained in http://stackoverflow.com/a/9338990, NSNotFound is
defined to be equal to NSIntegerMax, which is a 64-bit value in 64-bit
environment.  Hence we can not always store it into 'int' anymore,
because 'int' is always 32-bit value on OS X.

This CL touches only test code.  No user visible change is intended.

BUG=none
TEST=unittest
REF_BUG=3268453
REF_CL=105036313