Enable unittest for Linux and OSX on Travis-CI.
With this CL, each push / PR does is going to be tested with
unit tests on Travis-CI for Linux desktop and OSX build.

We don't plan to set up unit tests for Android and NaCl right now
because tests for them do not complete in host environment but
rather require special environments to run target binaries.