Hide mozc::base::Timer into mozc::base::Scheduler
Currently mozc::base::Timer is used only in mozc::base::Scheduler.
Hence it might be overkill to have mozc::base::Timer as a standalone
class.  With this CL, the same logic is embedded into the implementation
details of mozc::base::Scheduler.

This is just a code refactoring.  No behavior change is intended.

BUG=
TEST=
REF_BUG=
REF_CL=106275351