Deprecate EncodingUtil::UTF8ToSJIS.
This is a series of CLs to drop dependency on iconv (#252).

UTF8ToSJIS is used only in Windows build.  We can move it from
src/base/encoding_util.cc to src/win32/base/string_util.cc to in
favor of simplicity.

This is just a refactoring.  No behavior change is intended.

BUG=#252
TEST=unittest
REF_BUG=19010851
REF_CL=87124492