Fix #252: Make EncodingUtil platform-independent.
With this CL, EncodingUtil::SJISToUTF8 no longer depends on any
platform-dependent API, including iconv and Win32 APIs.

Keep in mind that EncodingUtil::SJISToUTF8 is mainly used for importing
user dictionary for MS-IME.  Hence we want EncodingUtil::SJISToUTF8 to
be compatible with the behavior on Windows.

This also closes #27 because we no longer depend on iconv.

BUG=#27,#252
TEST=unittest
REF_BUG=19010851
REF_CL=91370456,107291445