Build Mozc on Fedora 23 in Docker.
Since Fedora 21 will soon enter End Of Life (EOL) status on
December 1st, 2015, this CL updates Dockerfile for Fedora to use
Fedora 23 instead of Fedora 21.

Note that Android build support is intentionally dropped from
this Dockerfile due to the lack of due to the lack of OpenJDK 1.7
support in Fedora 23.

Note also that this Dockerfile is provided just for the reference.
Ensuring that Mozc can be built on Fedora 23 is beyond our goal
at the moment.

BUG=
TEST=compile and unittest in Docker 1.9.1 running on Ubuntu 14.04.3
REF_BUG=
REF_CL=