Stop using singleton of config::Config.
This refactoring solves more than five year long standing issues such
as:
 - The behavior of the module is unexpectedly changed due to the
   external change.
 - All unittests require to initialize and finalize the singleton,
   otherwise some unittest may violate environments of other unittests.
 - All unittests unnecessarily access file I/O and create files.

Background:

The config is a protocol buffer which stores user's various preferences
such as shortcut keys, preferred input style, usage stats opt-in flag,
etc.  The config is used by many modules in any layers.  So, we decided
to make it as a singleton and enabled all modules to read and write
config freely.  This decision actually accelerated the developmentd
velocity, when we were implementing the product from the
scratch. However, it became a technical debt for the sustainable
update.

In the current implementation, any modules can modify the contents of
config and it affects the results of other unittests.  So we need to
reset the status of config on the initialization and the finalization on
each unittest.  Even though some unittests are not affected by config,
it is unpredictable and team members including me tend to guard it just
in case.

  class FooTest : public ::testing::Test {
   protected:
    // Resets config not to be affected by other tests.
    virtual void SetUp() {
      Config config;
      ConfigHandler::GetDefaultConfig(&config);
      ConfigHandler::SetConfig(config);
    }

    // Resets config not to affect other tests.
    virtual void TearDown() {
      Config config;
      ConfigHandler::GetDefaultConfig(&config);
      ConfigHandler::SetConfig(config);
    }

Even worse, when we change the content of config via
ConfigHandler::SetConfig, the change is immediately updated to the file
storage.  It is reliable way to keep the consistency for client
applications, but not suitable for unittests.

This is obviously not the ideal way for code maintenance.  Thus, we will
stop using singleton of config and propagate it from upper layer to
lower layer.

No user-visible behavior change is intended.

BUG=
TEST=unittest
REF_BUG=19010851,19188911,22625214,23292110,23369771
REF_CL=94878367,97935879,99252909,101990432