Use OS_LINUX only for Linux desktop.
For historical reasons, OS_LINUX is a superset of OS_ANDROID and OS_NACL
and defined when building binaries for Android and NaCl.

With this CL, OS_LINUX is defined only when building Desktop linux
binaries.  Doing this makes it easier to consolidate platform-dependent
logics by preprocessor macros.  See the following table about how OS_*
macros are defined with this CL.

  target_platform |    host    |   target
  ------------------------------------------
         Android  |  OS_LINUX  | OS_ANDROID
            NaCl  |  OS_LINUX  | OS_NACL
           Linux  |       OS_LINUX
         Windows  |       OS_WIN
             Mac  |       OS_MACOSX

REF_BUG=21523433
REF_CL=97941571,97943618,98081675,98800504,105024909