Fix a runtime crash in Mozc for Android.
This is a follow up CL for 9cd4d36d2d51968bfad369fc6ee7703594007aef.

The root cause of the runtime crash is that base/base.gyp:jni_proxy is
not linked to libmozc.so only in OSS build because the link dependency
on 'jni_proxy' target is specified in 'net/net.gyp' only when
branding=="GoogleJapaneseInput".  There are two different problems.

 1. There are direct function invocations of base/android_jni_proxy.cc
    from android/jni/mozcjni.cc.  Hence it's 'android/android.gyp:mozc'
    that should have declared the dependency on 'base/base.gyp:jni_proxy'.
    Declaring the dependency in 'net/net.gyp' does not make much sense.
 2. With CL 96942303, 'base/base.gyp:jni_proxy' needs to be linked to
    libmozc.so even in OSS build.

This CL addresses both of avobe issues.

Closes #345.

BUG=#345
TEST=compile
REF_BUG=26369443
REF_CL=111290167