Update status icon even for Windows Store Apps.
On Windows 10 and later, Windows Store Apps can run in a window mode,
and IME status icons on the notification area are expected to be
updated.  This means that we have to initialize and update status icon
via LanguageBar API even for Windows Store Apps.

For the simplicity, with this CL, we start initializing and updating
status icons on LauguageBar even in immersive mode on Windows 8/8.1.
This can introduce some unnecessary overhead for those platforms but
basically no user-visible change should happen on Windows 8/8.1.

This CL also makes sure to update LanguageBar when
ITfThreadMgrEventSink::OnSetFocus is called.  This has not been obvious
in desktop apps but for Store Apps the lack of force update of
LanguageBar in ITfThreadMgrEventSink::OnSetFocus results in a stale icon
state on the language bar, which is so confusing.

Closes #350.

BUG=#350
TEST=manually done
REF_BUG=24593171,24793812
REF_CL=104530061,105271317,105271439,105279539,105283782