Developers should check out only master branch.
In Mozc repository, it's only master branch where we basically do not
change existing commit histories and commit hashes unless we
exceptionally have to do that.  Since any other branches are basically
used for development purpose, it is not recommended to check them out
unless needed.  Hence this CL updates documents and Dockerfile for
developers to not check them out accidentally.