Fix postbuilds_win.py --msvs_version handling
It turns out that '--msvs_version' option is not correctly passed to
postbuilds_win.py and it is always treated as the efault value 2013.

The root cause is a bug in postbuilds_win.gypi.  Since
'postbuild_msvs_option' GYP optoin is defined as an array, it should be
passed as '<@(postbuild_msvs_option)' instead of
'@(postbuild_msvs_option)'.  Otherwise, its value is never used.

With this CL, we can finally propagate '--msvs_version=2015' passed to
  build_mozc.py gyp
to postbuilds_win.py as expected.

BUG=#315
TEST=
REF_BUG=24654559
REF_CL=112532234
REF_TIME=2016-01-19T17:19:28-08:00
REF_TIME_RAW=1453252768 -0800