Replace CR+LF with LF in credits_*.html.
This CL does nothing but replace CR+LF with LF in credits_*.html.

The reason why we have used CR+LF there was actually for Internet
Explorer, which can recognize the Mark of the Web comment (MOTW) magic
comment only when the magic comment is followed by CR+LF.
https://msdn.microsoft.com/en-us/library/ms537628.aspx

A funny thing is that we decided to remove MOTW magic comment after
that, but forgot to revert back to use LF in those files.
Hence do this.  This is why we have used CR+LF there and wasted time
when resolving merge conflicts.

BUG=
TEST=
REF_BUG=
REF_CL=111825828