Introduce data set file and file embedding utility
The current dictionary, language model and relevant data set are mostly
embedded in executables. This approach has worked well on desktop, but
for mobile we want to have a clear and official boundary between logic
and code so that we can have more flexibility in terms of how we deploy
and update data set, because existing platform-specific hacks like
'use_separate_connection_data', 'use_separate_dictionary', and
PackedDataManager do not scale.

As the initial step, with this CL
 - Data set file packages multiple files into one file
 - embed_file.py utility embeds a file into C++ code

BUG=
TEST=
REF_BUG=26841123
REF_CL=113498663,113505577,113709549,128945999
REF_TIME=2016-02-01T16:04:17+09:00
REF_TIME_RAW=1454310257 +0900