Fix warning in HINSTANCE/uint32 conversion
|ShellExecuteW()| [1] is a tricky API in terms of how an error code is
encoded in the return value.  In short, if the return value is less than
or equal to 32, that is an error code.  Otherwise it means that the API
call succeeded.

This CL fixes the warning by casting the return value to size_t, which
is not ideal but should be better than the current code.

 [1]: https://msdn.microsoft.com/en-us/library/windows/desktop/bb762153.aspx

BUG=#315
TEST=
REF_BUG=26932823
REF_CL=114644224
REF_TIME=2016-02-14T09:03:31-08:00
REF_TIME_RAW=1455469411 -0800