Remove an incorrect conversion from HKL to LANGID
This CL fixes the following warning
  C4302: 'reinterpret_cast': truncation from 'HKL' to 'LANGID'
by removing the code in question.

This is actually an unintentional operation, which was introduced for a
long time ago.  KLID and HKL are two different concepts.  Converting one
with reinterpret_cast to the other is indeed an mistake.

That said, the impact on this code in question is supposed to be
limited, because it is only used for an early exit path.  The likelihood
of accidentally falling into that path is not so high.

BUG=
TEST=
REF_BUG=26932823
REF_CL=114656281
REF_TIME=2016-02-14T17:44:54-08:00
REF_TIME_RAW=1455500694 -0800