Fix warning around focus revision handlings
This CL addresses the following warnigns:
 - warning C4302:
    'reinterpret_cast': truncation from 'LPVOID' to 'int32'
 - warning C4311:
    'reinterpret_cast': pointer truncation from 'LPVOID' to 'int32'
 - warning C4312:
    'reinterpret_cast': conversion from 'const int32' to 'void *' of
    greater size

This CL also takes care of signed integer overflow, when incrementing
the focus revision, which is one of well-known undefined behaviour in
C++.

Note that in practice we do not depend on the order of the focus
revision in mozc_server.  Hence probably this has not been an issue
though unless the overflow results in runtime exceptions.

BUG=#315
TEST=
REF_BUG=26932823
REF_CL=114661183
REF_TIME=2016-02-14T20:31:43-08:00
REF_TIME_RAW=1455510703 -0800