Use CAN_USE_CANDIDATE_FORM_FOR_SUGGEST for Firefox.
Strictly speaking, IMM32 spec never requires the application to update
|CANDIDATEFORM| until IMN_OPENCANDIDATE is sent from the IME.

In the case of Mozc, we intentionally do not send IMN_OPENCANDIDATE when
the suggestion window is shown, because doing that can confuse some
applications such as command prompt.

That said, |CANDIDATEFORM::rcArea| is still a useful information that we
want to rely on if it is guaranteed to be kept updated.  For that
purpose, we have maintained a list of applications which are known to
keep |CANDIDATEFORM::rcArea| updated regardless of IMN_OPENCANDIDATE,
and this CL adds "MozillaWindowClass" window class into the whitelist so
that Firefox can be marked with CAN_USE_CANDIDATE_FORM_FOR_SUGGEST
compatibility flag.

BUG=#361
TEST=manually done
REF_BUG=27394949
REF_CL=115815738