Fix #374: Duplicate candidates after Undo
For all platforms, config.use_cascading_window is true by default.
For some platforms such as Android Session class does overwrite the
setting whether we actually show cascading window to ImeContext.

So the candidate list is not the same among the platforms.  Here, for
Windows/Mac, full-width "!" is not in the candidate list, but in the
cascading window.

BUG=#374
TEST=
REF_BUG=28285701
REF_CL=124801184,125129727
REF_TIME=2016-06-14T13:01:13+09:00
REF_TIME_RAW=1465876873 +0900