Rename 'mozc.imy' to 'mozc.data'
We decided to use *.imy as a hack because that was the only way to teach
Eclipse (to be precise Android SDK plugin for Eclipse) that those files
must not be compressed.

However, we no longer use Eclipse for Android development.  There is no
reason to keep using such an obscure file extension.

BUG=
TEST=
REF_BUG=14981159
REF_CL=125540611,127055505,130420025
REF_TIME=2016-06-22T17:07:55+09:00
REF_TIME_RAW=1466582875 +0900