Fix #378: Suppression word may not work
This is a corner case missed in the previous CL [1], which aimed to
reload the user dictionary only when necessary.

When the user dictionary reloader thread is not started, suppression
dictionary must be unlocked by main thread (Note: when the reloader
thread is started, the suppression dictionary is unlocked by it).

 [1]: b976132b2aa336c450e686a5cfc4602eb24c276b

BUG=#378
TEST=
REF_BUG=26499757,27639932
REF_CL=128960670
REF_TIME=2016-08-01T15:58:29+09:00
REF_TIME_RAW=1470034709 +0900