const void* overload of StringPiece::set must die
This mutator must be removed.  It is the only way to get a negative
length into a StringPiece without triggering an assert, it hides a
reinterpret_cast that should be visible at call sites, and it isn't
present in std::string_view.

BUG=
TEST=
REF_BUG=30675259
REF_CL=129656305
REF_TIME=2016-08-08T12:03:11-07:00
REF_TIME_RAW=1470682991 -0700