Teach 'build_mozc.py build' about Qt version
This is a preparation to switch to Qt5 on Windows.

Currently what option is specified to the configure option --qtver is
only available in 'build_mozc.py gyp' phase, and there is no way to know
it in 'build_mozc.py build', which makes it impossible to change the
behavior of 'build_mozc.py build' depending on --qtver settings.

With this CL, target Qt version specified to 'build_mozc.py gyp' is also
stored in mozc_version.txt so that we can use it later in 'build_mozc.py
build' phase, like other configuration options such as version code.

BUG=#327
TEST=
REF_BUG=26887740
REF_CL=131560502
REF_TIME=2016-08-28T23:35:45-07:00
REF_TIME_RAW=1472452545 -0700