Use the default font on Windows
We have specified Meiryo to GUI tools in Vista and later for better
look and feel.  However, at least on Windows 10, it doesn't look good
actually, and probably it'd be better to remove such a customization
in favor of simplicity.

BUG=
TEST=
REF_BUG=
REF_CL=140821194
REF_TIME=2016-12-02T16:58:04+09:00
REF_TIME_RAW=1480665484 +0900