#codehealth Get rid of one-shot keymap
It was possible to set one-shot keymap when sending SEND_KEY command,
but it looks like nobody uses this functionality.

One-shot keymap makes overlay keymap implementation complex so this CL removed it.

Now keymap is set to ImeContext when ImeContext.SetConfig() is called (== the configuration is explicitly set
as imposed or stored configuration).

PiperOrigin-RevId: 463211727