SessionHandler::GetStoredConfig() to set *merged* Config
GetStoredConfig() internally calls SetConfig() to ensure the on-memory config is the same as the locally stored one.

Before this CL, the parameter of SetConfig() was not *merged* but *stored* one so the behavior of the Sessions was unintentionally changed.

This CL fixes this behavior by setting *merged* config.

Note that this fix is effectively no-op because Gboard is the only client which uses imposed-config and doesn't use GetStoredConfig().

Note: Now SetConfig() always receives the same parameter (config::ConfigHandler::GetConfig()) so it's good time for refactoring.
PiperOrigin-RevId: 469128075