Introduce additioinal_character_group_capability into commands.proto, and enable/disable SingleHentaiganaRewriter based on the field.
The implementation is based on the discussion for cl/467132473.

This CL contains additional check for all candidates, which can be potentially harmful. I conduct session_handler_benchmark_test (iteration=100, random_seed=0) for before/after.

* before
 1. 76.9s (CPU time)
 2. 80.8s (CPU time)
 3. 76.4s (CPU time)

* after
 1. 78.8s (CPU time)
 2. 76.9s (CPU time)
 3. 77.8s (CPU time)

This result shows that the effect of performance caused by the introduction of Hentaigana filters is, at least, not very negative. There is no queue which indicates performance degradation.

PiperOrigin-RevId: 471448153