Unify superscript/subscript conversion into mixed conversion in SmallLetterRewriter
This CL allows to use ^ or _ in the middle of input. It enables conversion like "x^2+y^2=z^2" into "x²+y²=z²". By this addition, chemical formula like C_6H_12O_6 will be correctly converted into C₆H₁₂O₆.

This CL contains a small regression compared to current implementation. It was possible to input superscript/subscript for +-=() with prefix _/^ like "^1+1" to "¹⁺¹". While it is useful, this type of input creates ambiguity when it comes to input like "x^2+y^2=z^2". Because the latter will be more frequently inputted than the former, now superscript/subscript for these letters can be inputted only when directly prefixed by _/^. Therefore, "^1^+^1" is now valid input for "¹⁺¹".

PiperOrigin-RevId: 474432255