Clear undo context on Session::SetConfig()
SetRequest() and SetTable() have already cleared undo context.
To treat config in the same way as request, it's better to clear undo context also for SetConfig().

In reality, this is no-op because these setter methods are always called at the same time.
Therefore the undo context has already been cleared effectively when SetConfig() was called.

PiperOrigin-RevId: 478412465