Remove u8 string literal prefixes.
We already uses UTF-8 for source and execution encoding, so this prefix should be unnecessary and this change should be a no-op.

Additionally, in C++20, this prefix makes the value an array of char8_t*, not an array of char*, so most, if not all, of these usages don't build in C++20 with the prefix.

Also adds a couple of clang-format off/on blocks because it has trouble with emojis.

PiperOrigin-RevId: 479318075