Support overlay SessionKeymap
The overlay SessionKeymap is a collection of SessionKeymap which are loaded on top of the base SessionKeymap to apply partial modifications.

For example, when the base is MSIME, which maps `HENKAN` to `RECONVERT`, and an overlay maps `HENKAN` to `IME_ON`, the HENKAN key works as `IME_ON`.

This is for handy modification of keymapping. Desktop versions have already had the keymapping editor, but it's too much for mobile app. The overlays are expected to cover most of the requirements.

Startblock:
   rapid/google_keyboard tag Rapator exists
PiperOrigin-RevId: 508508373