Suppress redundant storage access of SetConfig
If previously stored config and the config about to be stored is identical (checked by hash),
skip storing it to reduce unneeded storage access.

This is important once we deprecate SET_IMPOSED_CONFIG, which stores the config on-memory.
By replacing it with SET_CONFIG, unneeded storage write-access will frequently happen
if we don't have this kind of optimization.

PiperOrigin-RevId: 510359387