Move ErrorToCanonicalCode and ErrorToCanonicalStatus to file_util to avoid cyclic dependency between file_util and win_util.
This is only problematic on Bazel and Win32 because each dependency and symbol is checked for each build target with Bazel, and file_util doesn't require win_util on other platforms.

These functions don't cover all error codes, and are only used in file_util, so they should be in the file_util file.

PiperOrigin-RevId: 511808356