Add Win32 library link targets to bazel/win32.
Use this target when a library/binary needs to link additional system libraries, such as aux_ulib.lib, ktmw32.lib, etc. Having this target in a non-Windows library is a no-op, so you don't need to put this inside mozc_select().

It's under the bazel directory because it's more like a linker option than library code.

PiperOrigin-RevId: 512381674