Stop supporting IMM32 Mozc in `UninstallHelper::GetNewEnabledProfileForVista()`.
We no longer need to consider the scenario for Windows XP/Vista/7 where IMM32 Mozc is installed on the machine.  This means that we can safely simplify `UninstallHelper` class.

This CL removes such a legacy handling code for Windows 7.

This CL also removes `UninstallHelperTest::BasicCaseForVista()` as it is not a supported scenario any more.

#codehealth

PiperOrigin-RevId: 516672000