Add TSF layout tests for `LayoutManager`
Unlike IMM32 Mozc, how `LayoutManager` works for TSF Mozc isn't that complicated.  This is because it is guaranteed that TSF Mozc sends `RendererCommand::Update` only when `ApplicationInfo::composition_target`
is available, and `ApplicationInfo::composition_target` is sufficient for the following methods to complete (except for the UI scaling factor).

- `LayoutManager::LayoutCandidateWindowForSuggestion`
- `LayoutManager::LayoutCandidateWindowForConversion`
- `LayoutManager::LayoutIndicatorWindow`

With above, this CL adds TSF layout test scenarios so that we can delete legacy test scenarios for IMM32 shortly.

#codehealth

PiperOrigin-RevId: 520203667