Use VS2019 for build instructions
Visual Studio 2017 15.3 and later do bundle `ninja.exe`, which means that we no longer need to ask developers to check out `depot_tools` for such a build environment.

To simplify build instructions, this CL updates the default value of `--msvs_version` option to be '2019' from '2017' with removing all the references to `depot_tools` from the build instructions.

#codehealth

PiperOrigin-RevId: 522758283