Use Visual Studio 2022 Community Edition for build instruction
While it is true that we can still build Mozc with Visual Studio 2017, let's use Visual Studio 2022 Community Edition for the build instruction for easier communication on the issue tracker.  Note that Visual Studio 2019 Community Edition is already difficult to install, which means that new contributors would be almost always using Visual Studio 2022.

#codehealth

PiperOrigin-RevId: 523308647