Normalize current working directory in `build_mozc.py` on Windows
It seems that the current working directory is not guaranteed to have the same uppercase letters and lowercase letters as the filesystem on Windows.  This is rather obvious on PowerShell, but technically it may happen in CI environment as well depending on how it's implemented and configured.

To make things deterministic and avoid confusions, let's use `build_mozc.py` to normalize the current working directory before invoking any command.

Fix #719

PiperOrigin-RevId: 523989676