Add ASSIGN_OR_RETURN_HRESULT and RETURN_IF_FAILED_HRESULT macros and HResult class.
ASSIGN_OR_RETURN_HRESULT is similar to ASSIGN_OR_RETURN for absl::StatusOr, but it's for HResultOr. It's a simpler implementation so it doesn't accept a comma (",") in lhs.

RETURN_IF_FAILED_HRESULT is similar to RETURN_IF_ERROR for absl::Status. It returns HRESULT if FAILED() returns true.

HResult is a helper class to construct HResultOr<T> without T.

PiperOrigin-RevId: 524550485