Add build rule for `omaha_util.cc` for OSS Mozc
This is a preparation to support building installers for Windows (#723).

In order to build custom action DLLs for OSS Mozc, we first need to be able to build `omaha_util.cc`, which has been disabled for OSS build configurations.

With this CL `OmahaUtil` can be built for OSS Mozc, while it remains to be no-op unless `build_mozc.py gyp --branding=GoogleJapaneseInput` is specified.

#codehealth

PiperOrigin-RevId: 525031670