Use `*.postbuild` file as a trigger to rebuild `*.msi` files
This is a preparation to support building installers for Windows (#723).

This CL addresses an existing bug that `build_mozc.py build package` may not correctly rebuild `*.msi` files for incremental build scenario.

The issue is that build target `generate_msi` has depended only on `wixobj_file`.  This means even if files like `mozc_server.exe` are updated, `generate_msi` target will not be triggered because `wixobj_file` remains to be the same.

This CL fixes such an issue by explicitly setting `*.postbuild` files as dependencies of `generate_msi`.

#codehealth

PiperOrigin-RevId: 525143755