Fix #729: Avoid redundant CI executions for forked repositories
This change makes the triggering rule for our default workflow a bit smarter for those who are going to fork or sync `google/mozc` to their repository on GitHub.

Previously, the build/test actions are triggered every time when they just fork or sync `google/mozc` to their repository on GitHub.  This is quite redundant because the same build/test are expected to be already triggered at `google/mozc`.

With this change, for forked repositories build/test actions will be triggered if and only if the branch is not the default branch. Hopefully this behavior would help contributors because they are most likely creating patches on some feature branches rather than the default branch.  Of course, people can freely customize workflow rules after forking `google/mozc`.

For `google/mozc` there must be no observable behavior change.

PiperOrigin-RevId: 525958867