Remove special condition to trigger GitHub Actions
This is a logical revert of 3b3d3bcfafd3a168804737d5d2b27a737d258094, which was made for #729.

It turns out that for public repositories there is no cost / resource implications for running GitHub Actions.

To simplify the workflow rules, let's not have any special condition to trigger GitHub Actions for now.

PiperOrigin-RevId: 526597280