Fix #730: Upload artifacts from GitHub Actions
With this CL, `mozc.zip` for Linux, `Mozc.pkg` for macOS, and `Mozc32.msi` and `Mozc64.msi` for Windows will be uploaded from GitHub Actions for each triggered run.

Anyone who sign in with GitHub account can download these artifacts up to retention period (90 days).  There is no cost/quota implications as far as the repository is a public repository.

https://docs.github.com/en/actions/using-workflows/storing-workflow-data-as-artifacts

Note that these binaries are not signed.

PiperOrigin-RevId: 526714757