Allow customizations of ComImplements with overrides and traits
Some of the COM classes in TIP requires a non-standard QueryInterface implementation. Some others derive from COM implementation classes. Make the IUnknown interface overridable to allow customizations.

This CL also introduces ComImplementsTraits. Currently, the only trait is OnObjectRelease, which is called when a COM object is released. This is necessary to prepare for unload in TIP.

PiperOrigin-RevId: 527599053