HResultOr<T>: use a custom base class instead of std::optional
This change saves four to eight bytes per HResult<T> instance depending on alignment.

std::optional has an additional bool flag, but this is unnecessary overhead as HResultOr<T> only has a value when ok() == true. This CL replaces std::optional with a set of custom base classes.
The semantics are the same as std::optional. Attributes like trivial copy / move / destructor and noexcept depend on the underlying type T.
PiperOrigin-RevId: 528360093