Support large IPC payload in Windows
It turns out that `win32_ipc.cc` may truncate IPC payload if it's bigger than `IPC_RESPONSESIZE`. It seems that Mozc's IPC implementations in other platforms don't have such a limitation, which is why #741 is observed only in Windows.

This CL takes advantage of so-called message-read mode in Windows NamedPipe.

The first step is to call `SetNamedPipeHandleState` API in the IPC client process so that the file handle returned from `CreateFile` can start behaving in the message-read mode.  This step is unnecessary in the IPC server process because `PIPE_READMODE_MESSAGE` is already set in `CreateNamedPipe`.

The next step is to carefully handle `ERROR_MORE_DATA`.  Basically we need to handle two more cases:

 - `ReadFile` returns `FALSE` with `ERROR_MORE_DATA`
 - `ReadFile` returns `FALSE` with `ERROR_IO_PENDING`, and `GetOverlappedResult` returns `FALSE` with `ERROR_MORE_DATA`

In both cases, we can continue reading the remaining data.

With this CL, `win32_ipc.cc` should be able to handle an arbitrary size of payloads as long as the memory allows.

Closes #741

PiperOrigin-RevId: 535319100