Fix #747: Quick fix for `ipc_test` in Linux
It seems that `GenerateInputData()` can be quite slow when generating a large random byte sequence.

As a quick workaround, let's move it a bit earlier so that its duration is not charged into IPC ping-pong operation.

In a subsequent CL I'll also try to remove the dependency on `mozc::Random` to speed it up and make it more predictable.

PiperOrigin-RevId: 535451763