Fix warning messages in GitHub Actions for Win
This is a follow up to e8d211e9f1293efc9026d9777dcd2e43687da0ed for #749.

Putting a line starts with `#` in `run` lines ends up with the following (non-critical) warning for `cmd` shell in Windows.  Let's move it outside of `run:` lines to keep the build log cleaner.

```
'#' is not recognized as an internal or external command,
operable program or batch file.
```

There should be no actual behavior change in the build process itself.

PiperOrigin-RevId: 536563466