Stop building 32-bit installer for Windows
With this CL, Windows build stops building 32-bit installer and executables used only for 32-bit Windows, which should reduce the overall build time because we need to build fewer number of executables.  This would also be an important step towards switching `mozc_tool.exe` from 32-bit to 64-bit as discussed in #732.

Note that there should be no user observable behavior changes in 64-bit edition of Windows, no matter whether the IME client process is built for 32-bit or 64-bit.

Closes #752.

#codehealth

PiperOrigin-RevId: 537221723