Migrates uses of associative containers to more performant types.
LSC doc: go/map-chooser-ga-lsc

This is a **NON-TRIVIAL AND PROBABLY SAFE (i.e. your tests passed)** change, not
a trivial one. We've identified the following migrations that apply for this
container:

*   The types in this container are small enough such that references to the
    underlying objects are unlikely to exist. Thus pointer stability is
    extraordinarily unlikely to be required, and will be migrated to a type
    without pointer stability guarantees.

**IF THIS CHANGE DOES NOT BUILD**, your targets are not registered with any TAP
builds (go/tap-blueprint).

These changes will save 100s of Terabytes of RAM and 100s-1000s of kGCUs and
also put us in a better place for responding to hash flooding attacks in the
future. These fall under the go/capacity-cy currently affecting our fleet.

We are collectively managing thousands of CLs for this effort. Please do not ask
us to fix pre-existing warnings. Reviewer-provided edits (go/edit-from-critique)
for nits are greatly appreciated as they save us a huge amount of time. Owners
are encouraged to change the types further in follow-up CLs.

MARKDOWN=true
#performance

Tested:
    TAP sample presubmit queue

PiperOrigin-RevId: 539087171