Wait for syncer for user data manager in SyncData
Background:
cl/15354340
`UserHistoryPredictor::Sync()` was made async so that DeleteSession command will not be blocked.

On the other hand, the client that issued SYNC_DATA command expect that the command should sync data synchronously.

Both of the implementation of DeletSession and SyncData commands call `UserHistoryPredictor::Sync()` internally.
So, we should call `UserHistoryPredictor::Wait()` in SyncData command.

PiperOrigin-RevId: 546156061