Make sure to exclude session_watch_dog_test from mac build
This is a follow up to 5ceb7a62ca36b4b1edcd6604aa38a7fa99c88125.

As mentioned in the document [1], --build_tag_filters doesn't affect
the set of tests executed with the 'test' command.  We need to use
'--test_tag_filters' instead.

With this commit 'session_watch_dog_test', which is known to be flaky
in mac environment, will be excluded from mac CI as intended [2].

 [1]: https://bazel.build/reference/command-line-reference#flag--build_tag_filters
 [2]: 9f36155e6f09d007a704983a588e66a5f6147e91

PiperOrigin-RevId: 552043275