Remove VerifyPrivilegeRestrictionIfNeeded from mozc_cache_service.cc
When we locked down mozc_cache_service with SERVICE_SID_TYPE_RESTRICTED,
one question raised between us and a security reviewer was how we could
verify that it's working as expected. After several discussions, we
added a special debug option --verify_privilege_restriction that's
available only for debug build to verify the behavior. For instance,

 * The following commands should succeed on Vista or later.
    sc stop MozcCacheService
    sc sidtype MozcCacheService restricted
    sc start MozcCacheService --verify_privilege_restriction

 * The following commands should fail on Vista or later.
    sc stop MozcCacheService
    sc sidtype MozcCacheService note
    sc start MozcCacheService --verify_privilege_restriction

This is dead code anyway. it'd probably make more sense to just remove
this in favor of simplicity.

There must be no observable behavior change anyway.

#codehealth

PiperOrigin-RevId: 552052031