Make 'build_mozc.py' compatible with Qt6 (#775)
With this commit 'build_mozc.py' starts passing 'qt_ver=6' to GYP as
needed.

For Windows and macOS, 'qt_ver=6' will be set when the Qt binaries
specified with '--qtdir' looks to be Qt6.

For Linux, 'qt_ver=6' will be set when the following command returns
exit code 0.

 pkg-config --exists Qt6Core Qt6Gui Qt6Widgets

There must be no behavior change in Bazel build.

PiperOrigin-RevId: 552727349