Fix no candidate window in Windows
This is a follow up to e7e8fb29b9bd671cc22cbe8efefdae95162dc722, which
accidentally introduced a deterministic access violation in
'mozc_renderer' in Windows.

The root cause is an unintentional behavior change in 'CreateFormat'.
This commit makes it clear that the buffer size needs to be carefully
specified.

Closes #788.

PiperOrigin-RevId: 554186217