Stop using shell=True in 'exec_command' in build_qt.py
This commit updates build_qt.py so that we can stop relying on
shell=True in 'exec_command' as suggested in the API doc [1].

This commit is basically about how to build Qt in Windows.
There must be no difference in the final artifacts.

This is a preparation of Qt6 migration in mac (#775).

#codehealth

 [1]: https://docs.python.org/3/library/subprocess.html#security-considerations

PiperOrigin-RevId: 555996264