Build intel64, arm64, and Universal Binary in GitHub Actions
With this commit, GitHub Actions for macOS starts building and uploading the
installer for the following 3 CPU architectures.

 - x86_64
 - arm64
 - universal (x86_64 and arm64)

One caveat is that so far we have not figured out how to build Qt6 for arm64 on
Intel mac.  As a result, Qt6 binaries in the arm64 installer are built as
universal binaries rather than a pure arm64 binaries.  This can be easily
fixed once GitHub start offering GitHub-hosted arm64 macOS runners.

In the build artifacts, it is clear that Universal Binary basically doubles the
installer size.

Closes #801.

PiperOrigin-RevId: 562468629