Remove unnecessary files from 'protoc_sources'
If a file is already included in 'protobuf_sources', there is no need to have
the same file in 'protoc_sources', as the 'protoc' target already depends on
'protobuf' target.

This is a mechanical cleanup for GYP build.  There must be no difference in the
build artifacts anyway.

This is a preparation for #803.

#codehealth

PiperOrigin-RevId: 563447681