Simplify TipRangeUtil::GetInputScopes a bit
This is a preparation for #362.

Retrieving an AppProperty from a given ITfRange as a VARIANT data is a
common task in TIP.  Let's extract out this operation from
TipRangeUtil::GetInputScopes so that we can reuse it in a subsequent
commit.

This is a mechanical change. There must be no behavior change.

PiperOrigin-RevId: 573457474