Exclude more unused features when building Qt6
As far as Qt binaries are concerned, we would prefer binary sizes over
performance as these binaries are not used so often.  This is rather
important for universal binary build in macOS, where we literally
bundle two binaries.

Excluding optional features is a widely-used way to reduce Qt binary
sizes [1], and Mozc has excluded a number of features already.

This commit aims to further expand the list of excluded features.
The overall results look good in both macOS and Windows.

 macOS aarch64 release build
  QtCore:         6354056 -> 5299752 = -1054304
  QtGui:          8237560 -> 7995160 =  -242400
  QtWidgets:      7136344 -> 6355144 =  -781200
                                       -2077904

 Windows x64_64 release build
  Qt6Core.dll:    5190656 -> 4316672 =  -873984
  Qt6Gui.dll:     4994048 -> 4694528 =  -299520
  Qt6Widgets.dll: 5074432 -> 4524544 =  -549888
                                       -1723392

It seems that excluding required features results in compile time error.
Thus there is expected to be no observable runtime issue.

Closes #822.

 [1]: https://doc.qt.io/qt-6/configure-options.html#including-or-excluding-features

PiperOrigin-RevId: 573824585