Automated g4 rollback of changelist 586135894.
*** Reason for rollback ***

Mozc's renderer_command.proto is shared between internal Bazel build and
external Bazel build [1].

The problem is that there seems to be no simple way to specify
'--experimental_editions' option in cc_proto_library in Bazel yet.

Thus renderer_command.proto cannot be converted to Edition until we
figure out a way to enable Edition in OSS Bazel build.

 [1]: https://github.com/google/mozc/blob/master/src/protocol/renderer_command.proto

*** Original change description ***

Migrate C++ protos to Editions (go/editions)

WARNING: We have determined that these protos are only used in C++.  If this is incorrect, or if you plan on using them in other languages within the next few months, a clean rollback should be safe.  Please notify us if you hit any issues.

This change has been verified to be behavior preserving both in the generated code and on the wire.  The only potential changes are for code that uses syntax-dependent reflection.  We've already migrated most of...

***

PiperOrigin-RevId: 586334075