Remove unused 'unicode name' data from emoji_data.tsv
This is already ignored data in the use-site, and eventually we can remove the column itself. For now, we just remove information in the column.
https://github.com/google/mozc/blob/master/src/rewriter/gen_emoji_rewriter_data.py

This is a part of cleanup for b/306504337. In the current impl of generate_emoji_data.py does not produce unicode name information, just replacing emoji_data.tsv with the generated new one will produce a lot of redundant diff. Another primary purpose of this CL is therefore reducing diff by that operation
https://github.com/google/mozc/blob/master/src/tools/emoji/generate_emoji_data.py

PiperOrigin-RevId: 588255880