Use `CCSIZEOF_STRUCT` in `win32_font_util.h`
This is a follow up to my previous commit [1], which introduced

  renderer/win32/win32_font_util.h

as part of our effort to remove the dependency on WTL (#861).

In that commit,

  NONCLIENTMETRICS::cbSize

was initialized with

  sizeof(NONCLIENTMETRICS),

which can cause a compatibility issue in the future if all the
following conditions are met [2]:

 * A future version of Windows introduces a new member to
   NONCLIENTMETRICS.
 * Mozc is built with WINVER that indicates such a new version of
   Windows.
 * Mozc binaries are running in Windows 10/11.

To eliminate such a possibility, we can use CCSIZEOF_STRUCT macro to
explicitly specify NONCLIENTMETRICS::cbSize as Windows Vista version.

 [1]: 3e428e35175ceb144fa741c6181d76ca28bcb4b4
 [2]: https://devblogs.microsoft.com/oldnewthing/20031212-00/?p=41523

#codehealth

PiperOrigin-RevId: 600498910