Clarify `candidate_window.cc` links to `msimg32.lib`
As part of our effort to remove the dependency on WTL (#861), this
commit makes it clear that candidate_window.cc needs to link to
msimg32.lib so that the following APIs can be used.

 * AlphaBlend
 * GradientFill

This dependency was not obvious as <atlgdi.h> provided in WTL
automatically adds it with a pragma directive, but we are going to
stop including <atlgdi.h> everywhere so this explicit dependency is
necessary.

#codehealth

PiperOrigin-RevId: 602499385