Stop using TxF in `FileUtil::AtomicRename`
Transactional NTFS (TxF) is now strongly discouraged and subject to be
removed in future versions of Windows [1]. The new file system called
Resilient File System (ReFS) has never been compatible with TxF [2].

To avoid unexpected compatibility issues in the future, let's drop the
dependency on TxF from FileUtil::AtomicRename with the existing
fallback logic, which we used to always rely on in Windows XP where TxF
was not available.

'FileUtilTest::AtomicRename' tests it in a comprehensive manner thus
there must be no user observable behavior change.

Closes #873.

#codehealth

 [1]: https://learn.microsoft.com/en-us/windows/win32/fileio/deprecation-of-txf
 [2]: https://learn.microsoft.com/en-us/windows-server/storage/refs/refs-overview#the-following-features-are-unavailable-on-refs-at-this-time

PiperOrigin-RevId: 604882971