Fix const-ness of `Segments` iterators
When the `Segments` has const-ness (e.g., `const Segments` or `const Segments&`), dereferencing its iterators (e.g., `*segments.begin()`) returned `Segment&`. This patch fixes it to `const Segment&`.

This needs to be explicitly handled because the `const_iterator` of the underlying storage the `std::deque<Segment *>` dereferences to non-const `Segment`.

This patch also makes `iterator` convertible to `const_iterator`. This is one of requirements in the STL container.

PiperOrigin-RevId: 608878528