Call absl::InitializeLog() before adding file sink.
This is not documented in https://abseil.io/docs/cpp/guides/logging nor discoverable in any way, but according to the function comment of `absl::InitializeLog` this is required.

That said, despite what the comment says additional log sink seems to work without this and the only effect seems to enable stderr thresholding and stack trace (from what I saw when reading source code).

This also suppresses `WARNING: All log messages before absl::InitializeLog() is called are written to STDERR` warning, which is nice :)

PiperOrigin-RevId: 611506844