Normalize quite in WiX rules
This is a preparation to switch to WiX v4 (#894).

While there is an automated converter from WiX v3 rules to WiX v4 rules,
it normalizes all the single quotes in XML attributes to double quotes.

To minimize the diffs in the above process, this commit applies the same
normalization in prior.

This is a mechanical replacement. There must be no difference in the
final installer files.

#codehealth

PiperOrigin-RevId: 614782079