Stop using `GYP_CROSSCOMPILE` in GYP builds
While GYP/ninja supports cross build by setting GYP_CROSSCOMPILE=1,
it was actually necessary only for the following scenarios, which
are all deprecated.

 * Build Mozc for Android on a Linux machine with GYP
 * Build Mozc for NaCl on a Linux machine with GYP
 * Build Mozc for iOS on a macOS machine with GYP

If the host environment and the target environment are exactly the
same, setting setting GYP_CROSSCOMPILE=1 is redundant and requires more
CPU cycles and disk space to complete the build.

Let's remove it as we no longer do cross build with GYP.

#codehealth

PiperOrigin-RevId: 615837660