Threshold as penalty refactoring for typing correction.
Before: score > threshold
After:  score - penalty > 0

We had been using threshold-based triggering control. i.e. trigger if score > threshold.

However, we recently introduced a conditional threshold in which the threshold is dynamically updated. This conditional threshold is confusing as we can update both the raw score and threshold. This refactoring fixes the threshold to be zero and controls the triggering condition only with the penalty. i.e, score - penalty > 0.0.  This CL migrates the thresholds to penalties. Since the scores are used in other ranking, this migration is flag guarded.

PiperOrigin-RevId: 636428033