Support `RewriteAd()` when the "年" suffix is at the start of the next segment
This patch allows `DateRewriter` rewriting without resizing when the next segment starts with the "年" suffix.

From the user perspective, when the `DateRewriter` rewrites Era to AD, the Era and the year must be in single segment, but the "年" suffix can be in the segment or at the start of the next segment.

The new behavior is consistent with `RewriteEra()`, which rewrites AD to Era. `RewriteEra()` supports when the next segment is equal to the "年" suffix, though it doesn't support when the "年" is in the year segment, or the next segment has following characters after the "年" suffix.

PiperOrigin-RevId: 642529056