Make .bazelrc compatible with msvc
Currently our .bazelrc specifies C/C++ options with an assumption that
the compiler recognizes gcc-like options, which of course do not work
when msvc is used. In order to support Bazel Windows build (#948),
it needs to be updated to specify copts based on the target environment.

There must be no observable behavior change in mac/Linux Bazel builds.

#codehealth

PiperOrigin-RevId: 645619204