Temporarily simplify magic header check in `*_data_manager.cc`
Historically we have passed the entire magic number of the embedded data
file to each *_data_manager.cc file so that it can compare the values
at run time. However the problem here is that propagating non-ASCII byte
data from the build system to *.cc file is not that simple task. This is
going to be more problematic when we start using Bazel on Windows, where
the backslash character has another special path separator meaning.

As a temporary solution to unblock Bazel migration (#948), this commit
relaxes the verification by only propagating the magic number byte
length instead of the entire byte array. Note that each data file still
has a checksum and the data integrity remains to be verifiable . The only
thing we can no longer verify is some accidental mistakes at the build
rules level, which is supposed to be unlikely to happen.

PiperOrigin-RevId: 646540433