Fix macOS builds due to typos in `.bazelrc`
This is a follow up commit to my previous commit [1], which aimed to add

  --noincompatible_enable_cc_toolchain_resolution

build option only to Windows builds as a preparation to support Bazel
build on Windows (#948). It, however, accidentally set the flag to mac
builds, which is why now macOS builds are failing.

With this commit the flag is set only to Windows builds as intended.

#codehealth

 [1]: 5efa371c639be5e06bdc797a9deef1dc53048f58

PiperOrigin-RevId: 646703582