Refactor `RewriteAd()` and `RewriteEra()` arguments
This patch moves `num_done` to an out argument to simplify the caller code.

It also changes `RewriteEra()` to be consistent with `RewriteAd()`, to skip rewriting the segment that starts with the `kNenKey`.

This makes the two functions more consistent, makes the caller a bit more readable, and reduces some minor CPU cycles. But the change isn't observable, unless we have a rewrite pattern that starts with the `kNenKey` in future, such as an era name starting with the `kNenKey`.

PiperOrigin-RevId: 650501471