Support long character completions with extra penalty.
In order to predict "ざいません" from "もうしわけご", we have to change the max_completion_length to 5, but long character completions would have large side-effect if the completion fails. We would like to add them if and only if the confident score is large enough.

The extra penalty is set as penalty = extended_per_char_penalty * (completed_length - max_char_length). When extend_per_char_penalty is zero, no penalty is added.

PiperOrigin-RevId: 662379828