Delete user history entry in rewriter in addition to predictor for DELETE_CANDIDATE_FROM_HISTORY command.
To fix the behavior, DeleteCandidate method was moved from UserDataManager to Converter. For now, the delete command is called only for the current segments.

Background:
Use history entry can be generated both for predictor and rewriter when we commit a candidate from mobile suggestion interaction.
However, DELETE_CANDIDATE_FROM_HISTORY command only delete entries for prediction, so the deleted candidate can be promoted for some cases.
PiperOrigin-RevId: 667466402