Enable switching to DIRECT mode using the mouse.
This is about addressing the problem where selecting DIRECT with the mouse to switch input modes doesn't work.

In the PropertyHandler::ProcessPropertyActivate function, the processing for DIRECT mode is being skipped. This is because DIRECT=0 as defined in src/protocol/commands.proto#L51, so !entry->composition_mode is evaluated as true.
As a result, while the DIRECT menu item exists, it is not linked to the actual processing that should be executed for DIRECT.