Build Win32 executables with required build config
As a preparation to use Bazel to build Mozc for Windows (#948), this
commit reworks my previous attempts [1][2], which aimed to make it
configurable on which executable should be built with what build
configurations (e.g. CRT link type, target CPU architecture).

For instance, the following command will build mozc_tool with dynamic
CRT for 64-bit, and mozc_tip32 with static CRT for 32-bit.

  bazel --bazelrc=windows.bazelrc build  \
      //win32/installer:mozc_tool        \
      //win32/installer:mozc_tip32       \
      --config oss_windows

There must be no observable behavior change in GYP build.

 [1]: 5efa371c639be5e06bdc797a9deef1dc53048f58
 [2]: ff649883e8e9809e016cbb6f135134b2366c0eb2

PiperOrigin-RevId: 680165643