#1002: Switch to bzlmod for Android build
This is a follow up to our previous commits [1][2], which enabled bzlmod
by default except for Android. With this commit, Android build actually
starts only bzlmod mode.

Basically the idea is to use transition [3] to specify required CPU and
platforms for each libmozc.so target. A bonus is that we no longer need
to rely on rules_android.

 [1]: e54fbcb776119ce2710f78d9e1767fb458876a61
 [2]: 699c2ce5b7d92476199e9a6ba1d83414c3558016
 [3]: https://bazel.build/rules/lib/builtins/transition

PiperOrigin-RevId: 680441153