Use the same fine names for installer when building with Bazel
This is another follow up commit to my previous commits [1], which aimed
to start supporting Mozc installer for Windows with Bazel (#948).

In GYP builds we used to have 64 suffix in the file names of Mozc
installers. Let's do the same thing for now to make our build system
transition simpler.

With this commit 'Mozc64_bazel.msi' finally starts showing up on the
GitHub Actions artifacts section.

 [1]: ffc460e0f3d487655ad52f320449a9acc6162b71
 [2]: b92c4fd2899f9027bf052c01031f429d41b38ef4

PiperOrigin-RevId: 680747948