Use `bazelisk` as a bazel launcher in Dockerfile
bazelisk [1] is a wrapper around bazel that allows you to use different
versions of bazel without having to install them all. This is useful
for testing different versions of bazel or for using a specific version
of bazel that is not available in the default repository.

Let's update Dockerfile to use bazelisk instead of bazel. This will
allow us to avoid a tricky question on which version of bazel should be
installed in the Docker container by checking in '.bazeliskrc' into the
repository.

Note that switching to recent versions of bazel is also necessary due to
our migration to bzlmod (#1002). Probably bazel 7.3 or later is
currently required to build Mozc with bazel.

 [1]: https://github.com/bazelbuild/bazelisk

#codehealth

PiperOrigin-RevId: 680750168